requirements.txt
streamlit==1.35.0
transformers==4.41.2
torch==2.3.0
Pillow==10.3.0
requests==2.32.2
